backend/requirements.txt
fastapi
uvicorn[standard]==0.24.0
python-multipart==0.0.6
ultralytics==8.3.235
opencv-python-headless==4.8.1.78
numpy==1.26.0  
Pillow==10.1.0
